Initial 'test' version of the broodROMLWP
In this early version 'broodDroid' moves over the screen every ~20 seconds.
Final goal is to make the broodDroid shoot while touching the screen.
This includes muzzle flashes and dropping bullets from the gun

So it becomes a little fun thing to make him shoot when he appears on the screen ;)